Change types.

+++ client/file_block.go
@@ -17,23 +17,23 @@ func (f *FileBlockDevice) Readonly() bool {
 	return false
 }
 
-func (f *FileBlockDevice) Size() uint64 {
+func (f *FileBlockDevice) Size() int64 {
 	info, err := f.file.Stat()
 	if err != nil {
 		log.Panicln("Error reading file info", err)
 	}
-	size := uint64(info.Size())
-	return size - (size % uint64(f.BlockSize()))
+	size := int64(info.Size())
+	return size - (size % int64(f.BlockSize()))
 }
 
-func (f *FileBlockDevice) BlockSize() uint32 {
+func (f *FileBlockDevice) BlockSize() int {
 	return 512
 }
 
 func (f *FileBlockDevice) ReadAt(p []byte, off int64) (n int, err error) {
 	if off%int64(f.BlockSize()) != 0 {
 		log.Panicln("Invalid offset", off)
-	} else if uint32(len(p))%f.BlockSize() != 0 {
+	} else if len(p)%f.BlockSize() != 0 {
 		log.Panicln("Invalid read length", len(p))
 	}
 	return f.file.ReadAt(p, off)
@@ -42,7 +42,7 @@ func (f *FileBlockDevice) ReadAt(p []byte, off int64) (n int, err error) {
 func (f *FileBlockDevice) WriteAt(p []byte, off int64) (n int, err error) {
 	if off%int64(f.BlockSize()) != 0 {
 		log.Panicln("Invalid offset", off)
-	} else if uint32(len(p))%f.BlockSize() != 0 {
+	} else if len(p)%f.BlockSize() != 0 {
 		log.Panicln("Invalid write length", len(p))
 	}
 	return f.file.WriteAt(p, off)
@@ -56,27 +56,27 @@ func (f *FileBlockDevice) Flush() error {
 	return f.file.Sync()
 }
 
-func (f *FileBlockDevice) Trim(off uint64, length uint32) error {
-	if off%uint64(f.BlockSize()) != 0 {
+func (f *FileBlockDevice) Trim(off int64, length uint32) error {
+	if off%int64(f.BlockSize()) != 0 {
 		log.Panicln("Invalid offset", off)
-	} else if length%f.BlockSize() != 0 {
+	} else if int(length)%f.BlockSize() != 0 {
 		log.Panicln("Invalid trim length", length)
 	}
 	// Writing zeros assumes the underlying filesystem can make the file sparse.
 	return f.writeZeros(off, length)
 }
 
-func (f *FileBlockDevice) writeZeros(off uint64, length uint32) error {
+func (f *FileBlockDevice) writeZeros(off int64, length uint32) error {
 	const oneMeg = 1024 * 1024
 	b := make([]byte, oneMeg)
-	for i := uint64(0); i < uint64(length); i += oneMeg {
+	for i := int64(0); i < int64(length); i += oneMeg {
 		start := off + i
 		end := off + i + oneMeg
-		if end > (off + uint64(length)) {
-			end = off + uint64(length)
+		if end > (off + int64(length)) {
+			end = off + int64(length)
 		}
 		l := end - start
-		_, err := f.file.WriteAt(b[:l], int64(start))
+		_, err := f.file.WriteAt(b[:l], start)
 		if err != nil {
 			return err
 		}
